fix: 답변 있는 질문 삭제 시 예외 처리 수정했습니다.

+++ src/main/java/com/coverflow/question/application/QuestionServiceImpl.java
@@ -149,6 +149,10 @@ public class QuestionServiceImpl implements QuestionService {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new QuestionNotFoundException(questionId));
 
+        if (!question.getAnswers().isEmpty()) {
+            throw new AnswerExistException(questionId);
+        }
+
         questionRepository.delete(question);
     }
 }
